moved duplicate logic into MemberInviteList component
